Remove duplicated code for txt2img mode
Remove duplicated code for txt2img mode in app_window.py by
replacing the code for the txt2img GUI with an instance of the
BaseWidget class.
